fix(build): 消除编译警告 - C4819 / C4267
- etest_core 添加 /EHsc 消除 spdlog 间接 include <ostream> 触发的 C4819
- 新增 cmake/libharu/patch_utf8.cmake，给 hpdf 目标加 /utf-8 消除 C4819
- LogHistoryBuffer.cpp 添加 static_cast<size_t> 消除有符号/无符号比较警告

+++ src/core/logger/LogHistoryBuffer.cpp
@@ -11,7 +11,7 @@ LogHistoryBuffer::~LogHistoryBuffer() = default;
 
 void LogHistoryBuffer::push(int level, const QString& text) {
   QMutexLocker locker(&mutex_);
-  if (buffer_.size() >= capacity_) {
+  if (buffer_.size() >= static_cast<size_t>(capacity_)) {
     buffer_.pop_front();
   }
   buffer_.push_back(LogEntry{level, text});
